refactor: mlfow setup is a function common to both services
